refactor to extract csrf functionality to separate module
