+++ src/attack.js
@@ -9,6 +9,7 @@ var logger  = require('./logger.js');
 // ================================================================================================
 
 function launch(config, onSuccess, onFail) {
+  var csrf  = require('./csrf.js')(config, onSuccess, onFail);
   var queue = async.queue(tryLogin, config.concurrency);
 
   queue.drain = function() {
@@ -34,57 +35,6 @@ function launch(config, onSuccess, onFail) {
   // Private 
   // ==============================================================================================
 
-  // Array of token data
-  var tokenPool = [];
-
-  // Get a CSRF token from the pool if available, otherwise request one
-  function getCSRF(callback) {
-    var sample = tokenPool.pop();
-
-    if (sample) {
-      callback(sample.token, sample.cookie);
-    } else {
-      requestCSRF(callback);
-    }
-  }
-
-  // Request CSRF token from server
-  function requestCSRF(callback) {
-    var reqConfig = config.getAdvanced();
-
-    request(reqConfig.csrf, function (error, response, body) {
-
-      if (!error && response.statusCode === 200) {
-        
-        collectCSRF(response, body, reqConfig.capture.csrfRegex, callback);
-
-      } else {
-        
-        logger.warn('Server responded with: ' + response.statusCode);
-        onFail();
-
-      }
-    });
-  }
-
-  // Collect CSRF token from response body and save to token pool
-  function collectCSRF(response, body, csrfRegex, callback) {
-
-    var cookieString = response.headers['set-cookie'] || '';
-    var capture      = body.match(new RegExp(csrfRegex));
-
-    if (!capture) {
-
-      logger.error('CSRF token not found');
-      logger.info('Debug info:\n');
-      logger.info(body);
-      onFail();
-      
-    }
-
-    callback(capture[1], cookieString);
-  }
-
   // Process server response to a login attempt
   function processLoginResponse(response, body, password) {
     
@@ -97,8 +47,8 @@ function launch(config, onSuccess, onFail) {
 
       logger.info('Invalid: ' + password);
       
-      collectCSRF(response, body, csrfRegex, function(token, cookieString) {
-        tokenPool.push({ 
+      csrf.collect(response, body, csrfRegex, function(token, cookieString) {
+        csrf.tokenPool.push({ 
           token: token, 
           cookie: cookieString 
         });
@@ -121,7 +71,7 @@ function launch(config, onSuccess, onFail) {
   // Attempt to login with given password
   function tryLogin(password, callback) {
 
-    getCSRF(function(token, cookieString) {
+    csrf.fetch(function(token, cookieString) {
       var opt = config.getLogin(password, token, cookieString);
 
       // Send a login POST
